requirements.txt
altair==4.2.2
attrs==23.1.0
blinker==1.6.2
boto3==1.26.133
botocore==1.29.133
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
clearml==1.10.4
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
filelock==3.12.0
fonttools==4.39.3
furl==2.1.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
orderedmultidict==1.0.1
packaging==23.1
pandas==2.0.1
pathlib2==2.3.7.post1
Pillow==9.5.0
proto-plus==1.22.2
protobuf==3.20.1
psutil==5.9.5
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.4.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.30.0
rich==13.3.5
rsa==4.9
s3transfer==0.6.1
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.22.2
six==1.16.0
smmap==5.0.0
streamlit==1.22.0
sympy==1.11.1
tenacity==8.2.2
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ultralytics==8.0.98
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
